SemanticData tests imports and model selection #348
- testing in progress
- fixed problem with null namespace application option
